Added handling of length of unicode dashes
The problem?

Our awesome box drawing was not working terribly well in docker
and the top and bottom lines were not lining up.

The solution!

Kept track of the number of dashes in each string before comparing
their lengths.  Once we subtracted out 2x the number of box-
drawing characters, we were left with the proper string lengths
and could comfortablyl line the two up using regular ol' math.