Added better boxing
The poblem?

Once we fixed box length calculation on docker, it broke under wsl <sigh>

The fix!

We added a check to make sure that the dash length was indeed 1.
When it is greater than 1, we recalculate the string lengths so
that we know the proper number of dashes to add.