[OOPS]  Fixed vim startup on windows since runtimepath does not include
~/.vim on windows

Problem?

When starting up, vim throws up an error message stating press enter to
continue.

Solution!

This problem was caused by an error in the vim rc that we found by using

```
:messages
```

When we reviewed the error, it had to do with vim not being able to find
the japanesque color file.  By adding `~/.vim` to the runtime path, that
cleared the error